requirements.txt
Pillow==2.2.1
matplotlib==2.0.2
numpy==1.13.3
torchvision==0.1.9
